+++ packages/common/src/index.js
@@ -62,6 +62,7 @@ export { getModalStyles } from './Utils/getModalStyles';
 export { mergeStyles } from './Utils/mergeStyles';
 export { default as normalizeInitialState } from './Utils/normalizeInitialState';
 export { validate } from './Utils/component-data-validator';
+export { isHexColor } from './Components/ColorPicker/utils';
 
 export {
   isValidUrl,